mpc85xx: Fix panic on SMP init via kernel patch
Initially, I found that multi-core mpc85xx boards like the P1020 will
panic upon initialization of the second CPU:

[    0.627233] smp: Bringing up secondary CPUs ...
[    0.681659] kernel tried to execute user page (0) - exploit attempt?
(uid: 0)
[    0.766618] BUG: Unable to handle kernel instruction fetch (NULL
pointer?)
[    0.848899] Faulting instruction address: 0x00000000
[    0.908273] Oops: Kernel access of bad area, sig: 11 [#1]

I bisected down to kernel commit 1023355234ca ("powerpc/85xx: fix
timebase sync issue when CONFIG_HOTPLUG_CPU=n"). I added in a
reversion to this commit as a patch to prevent the panic, but after
speaking with the original developer on the mailing list[1], I was
given a working solution, this patch.

[1]: https://lists.ozlabs.org/pipermail/linuxppc-dev/2021-November/237029.html